apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
strawberry-graphql[fastapi]==0.228.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
redis==5.0.4
python-multipart==0.0.9
resend==0.8.0
PyJWT==2.8.0
azure-communication-email==1.0.0
aiosmtplib==3.0.1
openpyxl==3.1.5
alembic==1.13.1
pytest==8.2.0
pytest-asyncio==0.23.6
aiosqlite==0.20.0
